requirements.txt
# Core API server
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pydantic>=2.6,<2.10
python-multipart>=0.0.9

# HTTP + environment
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

# AI generation
google-genai==1.68.0
openai>=1.30.0
Pillow>=10.0,<12.0

# Research / scraping
apify-client>=1.7.0
yt-dlp>=2024.8.6

# Supabase
supabase>=2.0.0

# Web search (Tavily — social platform search)
tavily-python>=0.5.0

# LinkedIn data parsing
pandas>=2.0.0

# Utility scripts (kept for full-repo shareability)
beautifulsoup4>=4.12,<5.0
filelock>=3.13.0
